pyproject.toml
[project]
name = "cinemabot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.22.0",
    "aiohttp>=3.12.15",
    "django>=5.2.5",
    "fastapi>=0.116.1",
    "python-dotenv>=1.1.1",
]
